#93 Commented out MiniMap's removeEntity method as well as the locations in which it is called in MoveAction and BaseWorld to stop the related RunTime exceptions from crashing the game when a dead unit is removed from the world.
Updates:
Made changes to the implimentation of Soldier's and Spacman's move away qualifiers to allow for units to be present on a tile with Water entities. The boolean method moveAway(List<BaseEntity> entities) was added to BaseEntity to help facilitate this. AI free movement still needs to be compensated for.
Flood DoT is now effective, however values need to be adjusted.
Floodable interface added and implemented in BuildingEntity to allow building's to know if they are under the flooding effect.
Added generic SpecialTileEntity class as an eventual replacement for Water (currently unused) and related enum.
Reverted threshold for onTick() to 10^5 from 10^7 until entity movement speed adjusted to correct values for 10^7.